switch to elementtree for XML parsing; implement an age table to only refresh news every 10 minutes

+++ mint/news.py
@@ -4,9 +4,38 @@
 # All Rights Reserved
 #
 
-import database
-import feedparser
+import time
+import urllib2 
+from elementtree import ElementTree
 
+import database
+
+REFRESH_TIME = 60 # seconds
+
+class NewsCacheAgeTable(database.DatabaseTable):
+    name = 'NewsCacheAge'
+    fields = ['age']
+
+    createSQL = "CREATE TABLE NewsCacheAge (age INT);"
+
+    def getAge(self):
+        cu = self.db.cursor()
+        cu.execute("SELECT age FROM NewsCacheAge")
+        
+        r = cu.fetchone()
+        if r:
+            age = r[0]
+        else:
+            age = 0
+        return age
+
+    def setAge(self):
+        cu = self.db.cursor()
+        cu.execute("DELETE FROM NewsCacheAge")
+        cu.execute("INSERT INTO NewsCacheAge VALUES (?)", time.time())
+        self.db.commit()
+        return True
+        
 class NewsCacheTable(database.KeyedTable):
     name = 'NewsCache'
     key = 'itemId'
@@ -17,28 +46,40 @@ class NewsCacheTable(database.KeyedTable):
                     content         STR,
                     link            STR,
                     category        STR
-                );"""
+                );
+                """
     fields = ['itemId', 'title', 'pubDate', 'content', 'link', 'category']
 
     def __init__(self, db, cfg):
         database.DatabaseTable.__init__(self, db)
         self.cfg = cfg
+        self.ageTable = NewsCacheAgeTable(db)
 
     def refresh(self, items = 5, purge = True):
         if not self.cfg.newsRssFeed:
-            return
+            return False
+
+        if (time.time() - REFRESH_TIME) < self.ageTable.getAge():
+            return False
 
         cu = self.db.cursor()
 
         if purge:
             cu.execute("DELETE FROM NewsCache")
 
-        fp = feedparser.parse(self.cfg.newsRssFeed)
+        url = urllib2.urlopen(self.cfg.newsRssFeed)
+        data = url.read()
 
-        for item in fp['items'][:5]:
+        tree = ElementTree.XML(data)
+        for item in tree.findall("channel/item")[:items]:
+            link = item.find("link").text
+            title = item.find("title").text
+            category = item.find("category").text
+        
             query = "INSERT INTO NewsCache VALUES (NULL, ?, ?, ?, ?, ?)"
-            cu.execute(query, item['title'], 0, item['content'][0]['value'][:60],
-                              item['link'], item['category'])
+            cu.execute(query, title, 0, '', link, category)
+        
+        self.ageTable.setAge()
         self.db.commit()
 
     def getNews(self):
@@ -53,4 +94,3 @@ class NewsCacheTable(database.KeyedTable):
                 item[key] = r[i]
             data.append(item)
         return data
-
